+++ layers/src/wt3dtfmnormloss_cuda.c
@@ -7,7 +7,7 @@
 extern THCState *state;
 
 // =============== FWD PASS ================== //
-int Weighted3DTransformNormLoss_forward_cuda(
+float Weighted3DTransformNormLoss_forward_cuda(
 			THCudaTensor *points,
 			THCudaTensor *masks,
 			THCudaTensor *tfms,
